remove old src/types.ts
