+++ src/types/index.ts
@@ -6,3 +6,4 @@ export * from "./flow";
 export * from "./sections";
 export * from "./session";
 export * from "./payment-request";
+export * from "./result";
